requirements.txt
asgiref==3.7.2
catboost==1.2.2
contourpy==1.1.1
cycler==0.11.0
Django==4.2.5
djangorestframework==3.14.0
fonttools==4.42.1
graphviz==0.20.1
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
plotly==5.17.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
